Fix ft_memset.c and ft_strmapi.c

+++ common_core/libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: gsantama <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/29 15:21:33 by gsantama          #+#    #+#             */
-/*   Updated: 2024/11/29 15:21:34 by gsantama         ###   ########.fr       */
+/*   Updated: 2024/11/30 18:03:16 by gsantama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < len)
